requirements.txt
google-cloud-texttospeech
pdfminer.six